requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.11.2
PyYAML==6.0
